refactor: modular architecture + repository cleanup (#5)
* refactor: modular architecture + repository cleanup (#3)

- mailer.ts (709行) → src/mailer/ (8ファイル)
- email.ts (473行) → src/email/ (5ファイル)
- SmtpTransportパターンで疎結合化
- parseCapabilities()蓄積バグ修正
- 公開API完全後方互換

Repository cleanup:
- CI/CD workflows: pnpm → Bun
- prettier.config.js 削除 (Biome移行済み)
- README.md 全面書き直し (全API文書化)
- README_zh-CN.md → README_ja.md (日本語版)
- CHANGELOG.md 更新 (v2.0.0)
- package.json リポジトリURL修正
- .npmignore: LICENSE + README_ja.md 追加
- .changeset/config.json: access→public
- transport.ts: 冗長な .toString() 削除

354 tests passing, Biome 0 errors.


* chore: update workers-types, @types/node, author field (#4)

- tsconfig.json: @cloudflare/workers-types/2023-07-01 → latest
- package.json: @types/node ^20.14.12 → ^22.15.0
- package.json: author zou-yu → ryyr-ry


---------
