+++ .github/workflows/publish.yml
@@ -2,7 +2,7 @@ name: Release & Publish
 
 on:
   push:
-    branches: [ main ]
+    branches: [main]
 
 jobs:
   release:
@@ -10,31 +10,29 @@ jobs:
     permissions:
       contents: write
       pull-requests: write
-    
+
     steps:
       - uses: actions/checkout@v4
         with:
           fetch-depth: 0
-      
+
+      - name: Setup Bun
+        uses: oven-sh/setup-bun@v2
+
       - name: Setup Node.js
         uses: actions/setup-node@v4
         with:
-          node-version: '20.x'
-          registry-url: 'https://registry.npmjs.org'
+          node-version: "22"
+          registry-url: "https://registry.npmjs.org"
 
-      - name: Setup pnpm
-        uses: pnpm/action-setup@v2
-        with:
-          version: latest
-          
       - name: Install dependencies
-        run: pnpm install --frozen-lockfile
-        
+        run: bun install --frozen-lockfile
+
       - name: Run tests
-        run: pnpm test
-        
+        run: bunx vitest run
+
       - name: Build
-        run: pnpm build
+        run: bun run build
 
       - name: Release
         env:
@@ -43,4 +41,4 @@ jobs:
           NPM_TOKEN: ${{ secrets.NPM_TOKEN }}
         uses: changesets/action@v1
         with:
-          publish: pnpm publish --no-git-checks
+          publish: bunx changeset publish
