+++ .npmignore
@@ -3,3 +3,5 @@
 !dist/**
 !package.json
 !README.md
+!README_ja.md
+!LICENSE
