--- prettier.config.js
@@ -1,11 +1,0 @@
-/** @type {import('prettier').Config} */
-module.exports = {
-  endOfLine: 'lf',
-  semi: false,
-  useTabs: false,
-  singleQuote: true,
-  arrowParens: 'avoid',
-  tabWidth: 2,
-  trailingComma: 'all',
-  bracketSpacing: true,
-}
