--- src/email.ts
@@ -1,473 +1,0 @@
-import type { SendResult } from "./result"
-import { arrayBufferToBase64, encode, encodeQuotedPrintable } from "./utils"
-import { validateEmail as checkEmail } from "./validate"
-
-function isAsciiOnly(text: string): boolean {
-	for (let i = 0; i < text.length; i++) {
-		const code = text.charCodeAt(i)
-		if (code >= 0x7f) return false
-		if (code <= 0x1f && code !== 0x09) return false
-	}
-	return true
-}
-
-export function encodeHeader(text: string): string {
-	if (isAsciiOnly(text)) {
-		return text
-	}
-	const bytes = encode(text)
-	const prefix = "=?UTF-8?Q?"
-	const suffix = "?="
-	const overhead = prefix.length + suffix.length
-	const maxPayload = 75 - overhead
-
-	const words: string[] = []
-	let current = ""
-
-	for (let i = 0; i < bytes.length; i++) {
-		const byte = bytes[i]
-		let fragment: string
-		if (byte >= 33 && byte <= 126 && byte !== 63 && byte !== 61 && byte !== 95) {
-			fragment = String.fromCharCode(byte)
-		} else if (byte === 32) {
-			fragment = "_"
-		} else {
-			fragment = `=${byte.toString(16).toUpperCase().padStart(2, "0")}`
-		}
-
-		const isUtf8LeadByte = (byte & 0xc0) !== 0x80
-		if (current.length + fragment.length > maxPayload && isUtf8LeadByte) {
-			words.push(`${prefix}${current}${suffix}`)
-			current = fragment
-		} else {
-			current += fragment
-		}
-	}
-	if (current.length > 0) {
-		words.push(`${prefix}${current}${suffix}`)
-	}
-
-	return words.join("\r\n ")
-}
-
-export type User = { name?: string; email: string }
-
-export type EmailOptions = {
-	from: string | User
-	to: string | string[] | User | User[]
-	reply?: string | User
-	cc?: string | string[] | User | User[]
-	bcc?: string | string[] | User | User[]
-	subject: string
-	text?: string
-	html?: string
-	headers?: Record<string, string>
-	attachments?: {
-		filename: string
-		content: string | Uint8Array | ArrayBuffer
-		mimeType?: string
-	}[]
-	dsnOverride?: {
-		envelopeId?: string
-		RET?: {
-			HEADERS?: boolean
-			FULL?: boolean
-		}
-		NOTIFY?: {
-			DELAY?: boolean
-			FAILURE?: boolean
-			SUCCESS?: boolean
-		}
-	}
-}
-
-export class Email {
-	public readonly from: User
-	public readonly to: User[]
-	public readonly reply?: User
-	public readonly cc?: User[]
-	public readonly bcc?: User[]
-
-	public readonly subject: string
-	public readonly text?: string
-	public readonly html?: string
-	public readonly dsnOverride?: {
-		envelopeId?: string
-		RET?: {
-			HEADERS?: boolean
-			FULL?: boolean
-		}
-		NOTIFY?: {
-			DELAY?: boolean
-			FAILURE?: boolean
-			SUCCESS?: boolean
-		}
-	}
-
-	public readonly attachments?: {
-		filename: string
-		content: string | Uint8Array | ArrayBuffer
-		mimeType?: string
-	}[]
-
-	public readonly headers: Record<string, string>
-
-	public setSentResult!: (result: SendResult) => void
-	public setSentError!: (e: unknown) => void
-
-	public sentResult = new Promise<SendResult>((resolve, reject) => {
-		this.setSentResult = resolve
-		this.setSentError = reject
-	})
-
-	public readonly sent: Promise<void> = this.sentResult.then(() => {})
-
-	public setSent(): void {
-		this.setSentResult({
-			messageId: this.headers["Message-ID"] ?? "",
-			accepted: [],
-			rejected: [],
-			responseTime: 0,
-			response: "",
-		})
-	}
-
-	constructor(options: EmailOptions) {
-		// sentResult が主要なエラーチャネルだが、sent のみ使用されるケースもあるため
-		// 両方の未処理拒否を抑制
-		this.sentResult.catch(() => {})
-		this.sent.catch(() => {})
-
-		if (!options.text && !options.html) {
-			throw new Error("At least one of text or html must be provided")
-		}
-
-		if (typeof options.from === "string") {
-			this.from = { email: options.from }
-		} else {
-			this.from = options.from
-		}
-		if (typeof options.reply === "string") {
-			this.reply = { email: options.reply }
-		} else {
-			this.reply = options.reply
-		}
-		this.to = Email.toUsers(options.to) ?? []
-		this.cc = Email.toUsers(options.cc)
-		this.bcc = Email.toUsers(options.bcc)
-
-		this.subject = options.subject
-		this.text = options.text
-		this.html = options.html
-		this.attachments = options.attachments
-		this.dsnOverride = options.dsnOverride
-		this.headers = options.headers || {}
-
-		this.validateNoCRLF()
-		this.validateEmailAddresses()
-		this.validateAttachments()
-	}
-
-	private static readonly CRLF_PATTERN = /[\r\n]/
-
-	private validateNoCRLF() {
-		if (Email.CRLF_PATTERN.test(this.subject)) {
-			throw new Error("CRLF injection detected in subject")
-		}
-		for (const [key, value] of Object.entries(this.headers)) {
-			if (Email.CRLF_PATTERN.test(key) || Email.CRLF_PATTERN.test(value)) {
-				throw new Error(`CRLF injection detected in header: ${key}`)
-			}
-		}
-	}
-
-	private validateEmailAddresses() {
-		this.validateEmail(this.from.email, "from")
-		for (const user of this.to) {
-			this.validateEmail(user.email, "to")
-		}
-		if (this.cc) {
-			for (const user of this.cc) {
-				this.validateEmail(user.email, "cc")
-			}
-		}
-		if (this.bcc) {
-			for (const user of this.bcc) {
-				this.validateEmail(user.email, "bcc")
-			}
-		}
-		if (this.reply) {
-			this.validateEmail(this.reply.email, "reply-to")
-		}
-	}
-
-	private static readonly ANGLE_BRACKET_PATTERN = /[<>]/
-
-	private validateEmail(email: string, field: string) {
-		if (Email.ANGLE_BRACKET_PATTERN.test(email)) {
-			throw new Error(`Invalid email address in ${field}: ${email}`)
-		}
-		const result = checkEmail(email)
-		if (!result.valid) {
-			throw new Error(`Invalid email address in ${field}: ${email}`)
-		}
-	}
-
-	private static readonly UNSAFE_FILENAME_PATTERN = /[\r\n"]/
-
-	private static readonly BASE64_PATTERN = /^[A-Za-z0-9+/\r\n]+=*$/
-
-	private validateAttachments() {
-		if (!this.attachments) return
-		for (const attachment of this.attachments) {
-			if (Email.UNSAFE_FILENAME_PATTERN.test(attachment.filename)) {
-				throw new Error(
-					`Invalid attachment filename: ${attachment.filename.replaceAll(/[\r\n]/g, "?")}`,
-				)
-			}
-			if (
-				typeof attachment.content === "string" &&
-				!Email.BASE64_PATTERN.test(attachment.content)
-			) {
-				throw new Error(`Invalid base64 content in attachment: ${attachment.filename}`)
-			}
-		}
-	}
-
-	private static foldHeaderLine(line: string, maxLen = 78): string {
-		if (line.includes("\r\n")) {
-			const segments = line.split("\r\n")
-			return segments.map((segment) => Email.foldHeaderLine(segment, maxLen)).join("\r\n")
-		}
-		if (line.length <= maxLen) return line
-
-		const parts: string[] = []
-		let remaining = line
-
-		while (remaining.length > maxLen) {
-			let splitAt = -1
-			for (let i = maxLen - 1; i > 0; i--) {
-				if (remaining[i] === " " || remaining[i] === "\t" || remaining[i] === ",") {
-					splitAt = remaining[i] === "," ? i + 1 : i
-					break
-				}
-			}
-			if (splitAt <= 0) {
-				splitAt = maxLen
-			}
-			parts.push(remaining.slice(0, splitAt))
-			remaining = remaining.slice(splitAt)
-			if (remaining.length > 0 && remaining[0] !== " " && remaining[0] !== "\t") {
-				remaining = ` ${remaining}`
-			}
-		}
-		if (remaining.length > 0) {
-			parts.push(remaining)
-		}
-		return parts.join("\r\n")
-	}
-
-	private static toUsers(user: string | string[] | User | User[] | undefined): User[] | undefined {
-		if (!user) {
-			return
-		}
-		if (typeof user === "string") {
-			return [{ email: user }]
-		} else if (Array.isArray(user)) {
-			return user.map((user) => {
-				if (typeof user === "string") {
-					return { email: user }
-				}
-				return user
-			})
-		} else {
-			return [user]
-		}
-	}
-
-	public getEmailData() {
-		this.resolveHeader()
-
-		const headersArray: string[] = ["MIME-Version: 1.0"]
-		for (const [key, value] of Object.entries(this.headers)) {
-			headersArray.push(Email.foldHeaderLine(`${key}: ${value}`))
-		}
-		const mixedBoundary = this.generateSafeBoundary("mixed_")
-		const alternativeBoundary = this.generateSafeBoundary("alternative_")
-
-		headersArray.push(
-			Email.foldHeaderLine(`Content-Type: multipart/mixed; boundary="${mixedBoundary}"`),
-		)
-		const headers = headersArray.join("\r\n")
-
-		let emailData = `${headers}\r\n\r\n`
-		emailData += `--${mixedBoundary}\r\n`
-
-		emailData += `Content-Type: multipart/alternative; boundary="${alternativeBoundary}"\r\n\r\n`
-
-		if (this.text) {
-			emailData += `--${alternativeBoundary}\r\n`
-			emailData += `Content-Type: text/plain; charset="UTF-8"\r\n`
-			emailData += `Content-Transfer-Encoding: quoted-printable\r\n\r\n`
-			const encodedText = encodeQuotedPrintable(this.text)
-			emailData += `${encodedText}\r\n\r\n`
-		}
-
-		if (this.html) {
-			emailData += `--${alternativeBoundary}\r\n`
-			emailData += `Content-Type: text/html; charset="UTF-8"\r\n`
-			emailData += `Content-Transfer-Encoding: quoted-printable\r\n\r\n`
-			const encodedHtml = encodeQuotedPrintable(this.html)
-			emailData += `${encodedHtml}\r\n\r\n`
-		}
-
-		emailData += `--${alternativeBoundary}--\r\n`
-
-		if (this.attachments) {
-			for (const attachment of this.attachments) {
-				const mimeType = attachment.mimeType || this.getMimeType(attachment.filename)
-				emailData += `--${mixedBoundary}\r\n`
-				emailData += `Content-Type: ${mimeType}; name="${attachment.filename}"\r\n`
-				emailData += `Content-Description: ${attachment.filename}\r\n`
-				emailData += `Content-Disposition: attachment; filename="${attachment.filename}"\r\n`
-				emailData += `Content-Transfer-Encoding: base64\r\n\r\n`
-
-				const base64Content =
-					typeof attachment.content === "string"
-						? attachment.content
-						: arrayBufferToBase64(
-								attachment.content instanceof ArrayBuffer
-									? attachment.content
-									: (attachment.content.buffer.slice(
-											attachment.content.byteOffset,
-											attachment.content.byteOffset + attachment.content.byteLength,
-										) as ArrayBuffer),
-							)
-
-				const lines = base64Content.match(/.{1,72}/g)
-				if (lines) {
-					emailData += `${lines.join("\r\n")}`
-				} else {
-					emailData += `${base64Content}`
-				}
-				emailData += "\r\n\r\n"
-			}
-		}
-		emailData += `--${mixedBoundary}--\r\n`
-
-		const safeEmailData = this.applyDotStuffing(emailData)
-
-		return `${safeEmailData}\r\n.\r\n`
-	}
-
-	private applyDotStuffing(data: string): string {
-		let result = data.replace(/\r\n\./g, "\r\n..")
-		if (result.startsWith(".")) {
-			result = `.${result}`
-		}
-		return result
-	}
-
-	private generateSafeBoundary(prefix: string): string {
-		const bytes = new Uint8Array(28)
-		crypto.getRandomValues(bytes)
-		const hex = Array.from(bytes)
-			.map((b) => b.toString(16).padStart(2, "0"))
-			.join("")
-
-		let boundary = prefix + hex
-		boundary = boundary.replace(/[<>@,;:\\/[\]?=" ]/g, "_") // Replace unwanted characters with '_'
-
-		return boundary
-	}
-
-	private getMimeType(filename: string): string {
-		const extension = filename.split(".").pop()?.toLowerCase()
-
-		const mimeTypes: { [key: string]: string } = {
-			txt: "text/plain",
-			html: "text/html",
-			csv: "text/csv",
-			pdf: "application/pdf",
-			png: "image/png",
-			jpg: "image/jpeg",
-			jpeg: "image/jpeg",
-			gif: "image/gif",
-			zip: "application/zip",
-		}
-
-		return mimeTypes[extension || "txt"] || "application/octet-stream" // Default to 'application/octet-stream'
-	}
-
-	private resolveHeader() {
-		this.resolveFrom()
-		this.resolveTo()
-		this.resolveReply()
-		this.resolveCC()
-		this.resolveSubject()
-		this.headers.Date = this.headers.Date ?? new Date().toUTCString()
-		this.headers["Message-ID"] =
-			this.headers["Message-ID"] ?? `<${crypto.randomUUID()}@${this.from.email.split("@").pop()}>`
-	}
-
-	private resolveFrom() {
-		if (this.headers.From) {
-			return
-		}
-		let from = this.from.email
-		if (this.from.name) {
-			from = `"${encodeHeader(this.from.name)}" <${from}>`
-		}
-		this.headers.From = from
-	}
-
-	private resolveTo() {
-		if (this.headers.To) {
-			return
-		}
-		const toAddresses = this.to.map((user) => {
-			if (user.name) {
-				return `"${encodeHeader(user.name)}" <${user.email}>`
-			}
-			return user.email
-		})
-		this.headers.To = toAddresses.join(", ")
-	}
-
-	private resolveSubject() {
-		if (this.headers.Subject) {
-			return
-		}
-		if (this.subject) {
-			this.headers.Subject = encodeHeader(this.subject)
-		}
-	}
-
-	private resolveReply() {
-		if (this.headers["Reply-To"]) {
-			return
-		}
-		if (this.reply) {
-			let replyAddress = this.reply.email
-			if (this.reply.name) {
-				replyAddress = `"${encodeHeader(this.reply.name)}" <${replyAddress}>`
-			}
-			this.headers["Reply-To"] = replyAddress
-		}
-	}
-
-	private resolveCC() {
-		if (this.headers.CC) {
-			return
-		}
-		if (this.cc) {
-			const ccAddresses = this.cc.map((user) => {
-				if (user.name) {
-					return `"${encodeHeader(user.name)}" <${user.email}>`
-				}
-				return user.email
-			})
-			this.headers.CC = ccAddresses.join(", ")
-		}
-	}
-}
