+++ src/email/index.ts
@@ -1,0 +1,4 @@
+export { Email } from "./email"
+export { encodeHeader, foldHeaderLine } from "./header"
+export { buildMimeMessage, resolveHeaders } from "./mime"
+export type { EmailOptions, User } from "./types"
