+++ src/mailer/index.ts
@@ -1,0 +1,10 @@
+export { WorkerMailerPool } from "./pool"
+export { SmtpTransport } from "./transport"
+export type {
+	AuthType,
+	Credentials,
+	DsnOptions,
+	SmtpCapabilities,
+	WorkerMailerOptions,
+} from "./types"
+export { WorkerMailer } from "./worker-mailer"
