+++ src/mailer/pool.ts
@@ -1,0 +1,45 @@
+import type { EmailOptions } from "../email"
+import type { SendResult } from "../result"
+import type { WorkerMailerOptions } from "./types"
+import { WorkerMailer } from "./worker-mailer"
+
+export class WorkerMailerPool {
+	private readonly options: WorkerMailerOptions
+	private readonly poolSize: number
+	private readonly mailers: WorkerMailer[] = []
+	private nextIndex = 0
+
+	constructor(options: WorkerMailerOptions & { poolSize?: number }) {
+		const { poolSize, ...mailerOptions } = options
+		this.poolSize = poolSize ?? 3
+		this.options = mailerOptions
+	}
+
+	async connect(): Promise<this> {
+		const connections = Array.from({ length: this.poolSize }, () =>
+			WorkerMailer.connect(this.options),
+		)
+		const mailers = await Promise.all(connections)
+		this.mailers.push(...mailers)
+		return this
+	}
+
+	send(options: EmailOptions): Promise<SendResult> {
+		if (this.mailers.length === 0) {
+			return Promise.reject(new Error("[WorkerMailerPool] Send failed: pool is not connected"))
+		}
+		const mailer = this.mailers[this.nextIndex % this.mailers.length]
+		this.nextIndex++
+		return mailer.send(options)
+	}
+
+	async close(): Promise<void> {
+		await Promise.all(this.mailers.map((mailer) => mailer.close()))
+		this.mailers.length = 0
+		this.nextIndex = 0
+	}
+
+	async [Symbol.asyncDispose](): Promise<void> {
+		await this.close()
+	}
+}
