+++ tsconfig.json
@@ -4,7 +4,7 @@
 		"lib": ["ESNext"],
 		"module": "ESNext",
 		"moduleResolution": "Bundler",
-		"types": ["@cloudflare/workers-types/2023-07-01", "@types/node"],
+		"types": ["@cloudflare/workers-types", "@types/node"],
 		"resolveJsonModule": true,
 		"checkJs": false,
 		"noEmit": true,
